+++ static/attraction.html
@@ -7,7 +7,7 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100..900&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="/views/layouts/attraction.css">
-  <script type="module" src="/views/scripts/attraction.js"></script>
+  <script type="module" src="/controllers/attractionController.js"></script>
   <title>Index</title>
 </head>
 <body>
